Cleaning up some leftovers from refactoring jena/hadoop/kryo
