+++ sansa-hadoop-jena/src/test/java/net/sansa_stack/hadoop/RecordReaderRdfTestBase.java
@@ -1,4 +1,4 @@
-package net.sansa_stack.io.hadoop;
+package net.sansa_stack.hadoop;
 
 import com.google.common.collect.*;
 import io.reactivex.rxjava3.core.Flowable;
@@ -9,7 +9,6 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.mapreduce.*;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.task.TaskAttemptContextImpl;
-import org.apache.jena.graph.Graph;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.query.DatasetFactory;
 import org.apache.jena.riot.RDFDataMgr;
@@ -25,7 +24,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
