+++ sansa-rdf/sansa-rdf-spark/src/main/scala/net/sansa_stack/rdf/spark/io/JenaKryoRegistrator.scala
@@ -1,10 +1,16 @@
 package net.sansa_stack.rdf.spark.io
 
 import com.esotericsoftware.kryo.Kryo
+import de.javakaffee.kryoserializers.guava.HashMultimapSerializer
+import net.sansa_stack.kryo.jena.JenaKryoRegistratorLib
 import org.apache.spark.serializer.KryoRegistrator
 
 /**
  * Created by nilesh on 01/06/2016.
+ *
+ * @author Nilesh
+ * @author Claus Stadler
+ * @author Lorenz Buehmann
  */
 class JenaKryoRegistrator extends KryoRegistrator {
   override def registerClasses(kryo: Kryo) {
